+++ src/graph/__init__.py
@@ -1,0 +1,1 @@
+from .geograph import GeoDigraph