+++ src/utils/serialization.py
@@ -3,6 +3,29 @@ import hashlib
 import pickle
 import os
 
+
+def load_checkpoint(ckpt_file_name, obj=None):
+    _dict = {}
+    if obj is not None and hasattr(obj, "__dict__"):
+        _dict = obj.__dict__
+
+    with open(ckpt_file_name,'rb') as f:
+        dict_ = pickle.load(f)
+    _dict.update(dict_)
+    
+    return _dict
+    
+
+def save_checkpoint(obj, ckpt_file_name, ignore_att=[]):
+    try:                
+        with open(ckpt_file_name, 'wb') as f:
+            pickle.dump({ k: v for k, v in obj.__dict__.items() if k not in ignore_att}, f)
+        return True
+    except:
+        return False
+
+
+
 class PickleSaver():
     def __init__(self, folder='../cache'):
         self.create_time=time.time()
@@ -69,9 +92,12 @@ class Saver:
 
 
 if __name__ == "__main__":
-    fn = "../../cache/tmp.pkl"
-    tmp = Saver(fn)
-    print(tmp.create_time)
-    tmp.save_()
+    # fn = "../../cache/tmp.pkl"
+    # tmp = Saver(fn)
+    # print(tmp.create_time)
+    # tmp.save_()
     
-    tmp.load_(fn)
+    # tmp.load_(fn)
+    
+    ckpt = '../../cache/Shenzhen.graph.ckpt'
+    info = load_checkpoint(ckpt)